portability changes and refactoring remove mpg123 and use pygame refactor tasmota devices control
