+++ Requirements/GassistPi-system-requirements.txt
@@ -10,4 +10,3 @@ libssl-dev
 python3-dev
 python3-venv
 socat
-mpg123
